+++ esbuild.js
@@ -1,13 +1,14 @@
-const { build } = require("esbuild")
+const { build } = require('esbuild')
 
 build({
-  entryPoints: ["./app.ts"],
+  entryPoints: ['./app.ts'],
   minify: true,
   keepNames: true,
   bundle: true,
   sourcemap: true,
-  platform: "node",
-  outfile: "./build/index.js"
+  platform: 'node',
+  external: ['pg-hstore'],
+  outfile: './build/index.js',
 }).catch(() => {
   process.exit(1)
 })