+++ CHANGELOG.md
@@ -1,11 +1,54 @@
 # Release Notes
 
+
+## v1.0.8 (2023-07-22)
+
+- table with index
+- backupDB split data with index
+
+## v1.0.7 (2023-07-22)
+
+- config/db.ts => environment variables
+
+## v1.0.6 (2023-07-22)
+
+- Could not resolve "pg-hstore" => error ketika build => external: ['pg-hstore'] => solved
+- https://stackoverflow.com/questions/73903039/typescript-serverless-esbuild-exclude-full-node-modules-doesnt-seem-to-wo
+
+## v1.0.5 (2023-07-21)
+
+- informasi setting nama database
+
+## v1.0.4 (2023-07-21)
+
+- penggunaan
+
+## v1.0.3 (2023-07-21)
+
+- npm install --save sequelize
+- npm install --save mysql2
+- get provinces
+- get filtered provinces
+- get filtered regency
+- get filtered district
+- get filtered village
+- get regency
+- get district
+- get village
+
+## v1.0.2 (2023-07-21)
+
+- controller
+- router
+
 ## v1.0.1 (2023-07-21)
 
 - npm i -D @types/bcryptjs @types/cors @types/express @types/jsonwebtoken
 - npm i -D esbuild
 - npm i -D ts-node-dev
 - npm i -D typescript
+- npm i -D @typescript-eslint/parser
+- npm i -D @typescript-eslint/eslint-plugin
 - npx tsc --init
 
 ## v1.0.0 (2023-07-07)